Формируется отчет "Заключение"
- Добавлены закладки в файл шаблона заключение
- Изменен класс IApartment - включен объект типа IObject
- Изменен класс IReport - включен объект типа IObject
- Добавлены поля цена, курс доллара и уценка в интерфейс IObject
- Добавлено поле площадь по СНиП в интерфейс IApartment
- Добавлены закладка с полями в форму "Квартира"

+++ Program/PRInterfaces/Interfaces/IObject.cs
@@ -42,17 +42,17 @@ namespace PRInterfaces.Interfaces
         /// <summary>
         /// Свойство. Задает и возвращает оценочную рыночную стоимость
         /// </summary>
-        int Price { get; set; }
+        float Price { get; set; }
 
         /// <summary>
         /// Свойство. Задает и возвращает скидку по ликвидационной стоимости
         /// </summary>
-        int Discount { get; set; }
+        float Discount { get; set; }
 
         /// <summary>
         /// Свойство. Задает и возвращает курс доллара
         /// </summary>
-        int Dollar { get; set; }
+        float Dollar { get; set; }
 
         /// <summary>
         /// Свойство. Задает и возвращает цель оценки
